pyproject.toml
[tool.poetry]
name = "asr-hw"
version = "0.1.0"
description = "My ASR homework"
authors = ["Artur Gimranov"]
readme = "README.md"

packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = ">=3.9"
torch = "==1.11.0"
torchvision = "==0.12.0"
torchaudio = "==0.11.0"
hydra-core = "==1.3"
kenlm = "==0.2.0"
pyctcdecode = "==0.5.0"
numpy = "*"
tqdm = "*"
tensorboard = "*"
matplotlib = "*"
pandas = "*"
editdistance = "*"
wandb = "*"
datasets = "*"
speechbrain = "*"
torch_audiomentations = "*"
gdown = "*"
wget = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
